refactoring of logging logic wip
