+++ ZipUtility/Logging/Common/ILogEntryWriter.cs
@@ -1,0 +1,11 @@
+﻿using SRF.FileLogging.Models;
+
+namespace SRF.FileLogging.Common
+{
+  public interface ILogEntryWriter
+  {
+    string GenerateLogHeaderLine();
+
+    string GenerateLogEntryLine(LogEntry Info);
+  }
+}
